Rename: manager entity directory
