+++ src/main/java/world/trecord/domain/feedcontributor/ManagerRepository.java
@@ -1,4 +1,4 @@
-package world.trecord.domain.manager;
+package world.trecord.domain.feedcontributor;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
