+++ src/main/java/world/trecord/service/invitation/InvitationService.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import world.trecord.domain.feed.FeedEntity;
 import world.trecord.domain.feed.FeedRepository;
+import world.trecord.domain.feedcontributor.ManagerEntity;
+import world.trecord.domain.feedcontributor.ManagerRepository;
 import world.trecord.domain.invitation.InvitationEntity;
 import world.trecord.domain.invitation.InvitationRepository;
-import world.trecord.domain.manager.ManagerEntity;
-import world.trecord.domain.manager.ManagerRepository;
 import world.trecord.domain.notification.args.NotificationArgs;
 import world.trecord.domain.users.UserEntity;
 import world.trecord.domain.users.UserRepository;
