+++ src/test/java/world/trecord/service/invitation/InvitationServiceConcurrencyTest.java
@@ -8,8 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import world.trecord.domain.feed.FeedEntity;
 import world.trecord.domain.feed.FeedRepository;
+import world.trecord.domain.feedcontributor.ManagerRepository;
 import world.trecord.domain.invitation.InvitationRepository;
-import world.trecord.domain.manager.ManagerRepository;
 import world.trecord.domain.users.UserEntity;
 import world.trecord.domain.users.UserRepository;
 import world.trecord.event.notification.NotificationEventListener;
