+++ src/test/java/world/trecord/service/invitation/InvitationServiceTest.java
@@ -10,11 +10,11 @@ import org.springframework.transaction.annotation.Transactional;
 import org.testcontainers.shaded.org.awaitility.Awaitility;
 import world.trecord.domain.feed.FeedEntity;
 import world.trecord.domain.feed.FeedRepository;
+import world.trecord.domain.feedcontributor.ManagerEntity;
+import world.trecord.domain.feedcontributor.ManagerPermission;
+import world.trecord.domain.feedcontributor.ManagerRepository;
 import world.trecord.domain.invitation.InvitationRepository;
 import world.trecord.domain.invitation.InvitationStatus;
-import world.trecord.domain.manager.ManagerEntity;
-import world.trecord.domain.manager.ManagerPermission;
-import world.trecord.domain.manager.ManagerRepository;
 import world.trecord.domain.users.UserEntity;
 import world.trecord.domain.users.UserRepository;
 import world.trecord.event.notification.NotificationEventListener;
